+++ ingenialink/canopen/network.py
@@ -23,6 +23,7 @@ from ingenialink.canopen.servo import CANOPEN_SDO_RESPONSE_TIMEOUT, CanopenServo
 from ingenialink.enums.register import RegAccess, RegCyclicType, RegDtype
 from ingenialink.exceptions import ILError, ILFirmwareLoadError
 from ingenialink.network import NetDevEvt, NetProt, NetState, Network, SlaveInfo
+from ingenialink.servo import Servo
 from ingenialink.utils._utils import DisableLogger, convert_bytes_to_dtype
 from ingenialink.utils.mcb import MCB
 
@@ -347,6 +348,7 @@ class CanopenNetwork(Network):
         dictionary: str,
         servo_status_listener: bool = False,
         net_status_listener: bool = False,
+        disconnect_callback: Optional[Callable[[Servo], None]] = None,
     ) -> CanopenServo:
         """Connects to a drive through a given target node ID.
 
@@ -357,6 +359,8 @@ class CanopenNetwork(Network):
                 its status, errors, faults, etc.
             net_status_listener: Toggle the listener of the network
                 status, connection and disconnection.
+            disconnect_callback: Callback function to be called when the servo is disconnected.
+                If not specified, no callback will be called.
 
         Raises:
             ILError: if there aren't nodes in the network.
@@ -382,7 +386,11 @@ class CanopenNetwork(Network):
                 node.nmt.start_node_guarding(self.NODE_GUARDING_PERIOD_S)
 
                 servo = CanopenServo(
-                    target, node, dictionary, servo_status_listener=servo_status_listener
+                    target,
+                    node,
+                    dictionary,
+                    servo_status_listener=servo_status_listener,
+                    disconnect_callback=disconnect_callback,
                 )
                 self.servos.append(servo)
                 self._set_servo_state(target, NetState.CONNECTED)
@@ -407,6 +415,9 @@ class CanopenNetwork(Network):
             servo: Instance of the servo connected.
 
         """
+        # Notify that disconnect_from_slave has been called
+        if servo._disconnect_callback:
+            servo._disconnect_callback(servo)
         self.stop_status_listener()
         servo.stop_status_listener()
         self.servos.remove(servo)