+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Specific methods for complete-access
 - Methods to read RPDOs and TPDOs from the ethercat slave.
 - Methods to get mapping registers in a PDOMap and respective mapped items.
+- Callback mechanism to notify when a servo is disconnected from a network.
 
 ### Fixed
 - Default xdf v2 injected safety registers